+++ frontend/journey-app/src/components/ParksPage/ParkCodes.jsx
@@ -1,0 +1,68 @@
+// parkCodes.js
+// This is a list of all the park codes and their corresponding park names.
+const parkCodes = {
+  "acad": "Acadia National Park",
+  "arch": "Arches National Park",
+  "bibe": "Big Bend National Park",
+  "blca": "Black Canyon of the Gunnison National Park",
+  "brca": "Bryce Canyon National Park",
+  "cany": "Canyonlands National Park",
+  "care": "Carlsbad Caverns National Park",
+  "cave": "Caverns of Sonora",
+  "chis": "Channel Islands National Park",
+  "cong": "Congaree National Park",
+  "crla": "Crater Lake National Park",
+  "cuva": "Cuyahoga Valley National Park",
+  "dena": "Denali National Park and Preserve",
+  "deva": "Death Valley National Park",
+  "drto": "Dry Tortugas National Park",
+  "ever": "Everglades National Park",
+  "fiis": "Fire Island National Seashore",
+  "flfo": "Florissant Fossil Beds National Monument",
+  "gaar": "Gates of the Arctic National Park and Preserve",
+  "glac": "Glacier National Park",
+  "glba": "Glacier Bay National Park and Preserve",
+  "grca": "Grand Canyon National Park",
+  "grsa": "Great Sand Dunes National Park and Preserve",
+  "grsm": "Great Smoky Mountains National Park",
+  "grte": "Grand Teton National Park",
+  "guco": "Guadalupe Mountains National Park",
+  "gumo": "Guadalupe Mountains National Park",
+  "hale": "Haleakalā National Park",
+  "havo": "Hawai'i Volcanoes National Park",
+  "hosp": "Hot Springs National Park",
+  "indu": "Indiana Dunes National Park",
+  "isro": "Isle Royale National Park",
+  "jotr": "Joshua Tree National Park",
+  "katm": "Katmai National Park and Preserve",
+  "kefj": "Kenai Fjords National Park",
+  "kica": "Kings Canyon National Park",
+  "klse": "Klondike Gold Rush National Historical Park",
+  "kova": "Kobuk Valley National Park",
+  "lacl": "Lake Clark National Park and Preserve",
+  "lavo": "Lassen Volcanic National Park",
+  "maca": "Mammoth Cave National Park",
+  "meve": "Mesa Verde National Park",
+  "mora": "Mount Rainier National Park",
+  "noca": "North Cascades National Park",
+  "olym": "Olympic National Park",
+  "pefo": "Petrified Forest National Park",
+  "pinn": "Pinnacles National Park",
+  "redw": "Redwood National and State Parks",
+  "romo": "Rocky Mountain National Park",
+  "sagu": "Saguaro National Park",
+  "seki": "Sequoia National Park",
+  "shen": "Shenandoah National Park",
+  "thro": "Theodore Roosevelt National Park",
+  "viis": "Virgin Islands National Park",
+  "voya": "Voyageurs National Park",
+  "whsa": "White Sands National Park",
+  "wica": "Wind Cave National Park",
+  "wrst": "Wrangell-St. Elias National Park and Preserve",
+  "Yellowstone National Park": "yell",
+  "yose": "Yosemite National Park",
+  "zion": "Zion National Park"
+}
+
+
+export default parkCodes;
