+++ frontend/journey-app/src/components/ParksPage/Videos.jsx
@@ -1,21 +1,21 @@
 import React, { useState, useEffect } from 'react';
 import axios from 'axios';
-import './Videos.css'
+import ParkCodes from './ParkCodes';
+import './Videos.css';
 
 const Videos = () => {
   const [videos, setVideos] = useState([]);
   const [originalVideos, setOriginalVideos] = useState([]);
-  const [searchTerm, setSearchTerm] = useState('');
+  const [selectedPark, setSelectedPark] = useState('');
 
   useEffect(() => {
     const fetchVideos = async () => {
       try {
-        const token = localStorage.getItem("token");
+        const token = localStorage.getItem('token');
         const response = await axios.get(`http://${import.meta.env.VITE_BASE_URL}/api/videos/`, {
           headers: { Authorization: `Token ${token}` },
         });
-        console.log("Videos response:", response);
-        setVideos(response.data.data);
+        console.log('Videos response:', response);
         setOriginalVideos(response.data.data);
       } catch (error) {
         console.error('Error fetching videos:', error);
@@ -25,42 +25,37 @@ const Videos = () => {
     fetchVideos();
   }, []);
 
-  const handleSearch = () => {
+  const handleParkChange = (e) => {
+    const parkName = e.target.value;
+    setSelectedPark(parkName);
     const filteredVideos = originalVideos.filter((video) =>
-      video.tags.some((tag) => tag.toLowerCase().includes(searchTerm.toLowerCase()))
+      video.tags.some((tag) => tag.toLowerCase().includes(parkName.toLowerCase()))
     );
     setVideos(filteredVideos);
-  };
-
-
-  const handleReset = () => {
-    setSearchTerm('');
-    setVideos(originalVideos);
   };
 
   return (
     <div className="videos-container">
       <h2 className="videos-header">Videos</h2>
       <div>
-        <input
-          type="text"
-          placeholder="Search by tag"
-          value={searchTerm}
-          onChange={(event) => setSearchTerm(event.target.value)}
-        />
-        <button className="button-small" onClick={handleSearch}>Search</button>
-        <button className="button-small" onClick={handleReset}>Reset</button>
+        <label htmlFor="parkFilter">Select a park: </label>
+        <select id="parkFilter" value={selectedPark} onChange={handleParkChange}>
+          <option value="">Select a park</option>
+          {Object.entries(ParkCodes).map(([code, name]) => (
+            <option key={code} value={name}>
+              {name}
+            </option>
+          ))}
+        </select>
       </div>
       <div className="videos-grid">
         {videos.map((video) => (
           <div key={video.id} className="video-card">
             <h3>{video.title}</h3>
             <video width="320" height="240" controls>
-              {video.versions.map((version) => (
-                version.heightPixels === 720 && (
-                  <source src={version.url} type={version.fileType} />
-                )
-              ))}
+              {video.versions.map((version) =>
+                version.heightPixels === 720 && <source src={version.url} type={version.fileType} />
+              )}
               Your browser does not support the video tag.
             </video>
             <p>{video.description}</p>
@@ -75,6 +70,11 @@ export default Videos;
 
 
 
+
+// import React, { useState, useEffect } from 'react';
+// import axios from 'axios';
+// import './Videos.css'
+
 // const Videos = () => {
 //   const [videos, setVideos] = useState([]);
 //   const [originalVideos, setOriginalVideos] = useState([]);
@@ -84,7 +84,7 @@ export default Videos;
 //     const fetchVideos = async () => {
 //       try {
 //         const token = localStorage.getItem("token");
-//         const response = await axios.get("http://localhost:8000/api/videos/", {
+//         const response = await axios.get(`http://${import.meta.env.VITE_BASE_URL}/api/videos/`, {
 //           headers: { Authorization: `Token ${token}` },
 //         });
 //         console.log("Videos response:", response);
